Misc updates including new abstraction layers
Misc updates triggered by some recent build failures. The
most substantial change is creating a new abstraction layer
for the 3-tier model. This allows me to move a fair number
of interfaces from the 'model' project to the respective layer.

These new layers will also contain tests that are implementation-
agnostic. I've used this approach in other projects - with the
simplest approach each kata needs to "implement" each test but
only to the extent of providing the correct subdirectory to scan
for the tested implementations. I have no doubt there's a cleaner
way to handle this with a custom test runner.

(Or the 'dojo' class could handle this if we're willing to break
the normal encapulation rules - we could create a Spring Profile
for each kata, with the corresponding jars marked as 'optional'.
I haven't looked at this in depth because this might become a
build/dependency hell.)